Re-Organised answer box handler files #67
